[Change] Floor plan models
Exchanged the prototype floor plan models with their final versions. Walls are fixed now so one can´t see the sky when looking through them anymore but one sees a 2nd and 3rd row of walls to make it seem like one sees the inside of the walls. Also edited some planes for the rooms and the posoition of the buttons for the elevator on the ground floor. Also deleted the outdated versions of room models.